extract workflow decision routes
